requirements.txt
pandas
numpy
plotly
scipy
streamlit>=1.32
joblib
scikit-learn
xgboost
altair<5
google-generativeai
python-dotenv